backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.135.1
GeoAlchemy2==0.18.4
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
packaging==26.0
passlib==1.7.4
Pillow==11.2.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.52.1
supabase==2.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.10.2
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
